+++ tests/test_metadata_api.py
@@ -786,7 +786,7 @@ class TestMockedMetadataSession:
         msg = "Missing required argument. One of the following args are required: oclcNumber, issn, isbn"
         with MetadataSession(authorization=mock_token) as session:
             with pytest.raises(WorldcatSessionError) as exc:
-                session.search_general_holdings(heldBy="NY", limit=20)
+                session.search_general_holdings(holdingsAllEditions=True, limit=20)
             assert msg in str(exc.value)
 
     def test_search_general_holdings_invalid_oclc_number(self, mock_token):
@@ -1022,6 +1022,7 @@ class TestLiveMetadataSession:
                 ]
             )
 
+    @pytest.mark.holdings
     def test_holding_set_unset(self, live_keys):
         token = WorldcatAccessToken(
             key=os.getenv("WCKey"),
@@ -1078,6 +1079,7 @@ class TestLiveMetadataSession:
                 "detail": None,
             }
 
+    @pytest.mark.holdings
     def test_holdings_set(self, live_keys):
         token = WorldcatAccessToken(
             key=os.getenv("WCKey"),
@@ -1111,6 +1113,7 @@ class TestLiveMetadataSession:
                 ]
             )
 
+    @pytest.mark.holdings
     def test_holdings_unset(self, live_keys):
         token = WorldcatAccessToken(
             key=os.getenv("WCKey"),